refactor: 딜러 클래스 추가 및 GameParticipant로 추상화
