+++ src/main/java/domain/Player.java
@@ -1,14 +1,9 @@
 package domain;
 
-import java.util.Set;
-
-public class Player {
-
-    public static final int SUM_LIMIT = 21;
-    private final Cards cards;
+public class Player extends GameParticipant {
 
     private Player(Cards cards) {
-        this.cards = cards;
+        super(cards);
     }
 
     public static Player init() {
@@ -17,41 +12,5 @@ public class Player {
 
     public static Player of(Cards cards) {
         return new Player(cards);
-    }
-
-    public void addCard(Card card) {
-        cards.add(card);
-    }
-
-    public Cards getCards() {
-        return cards;
-    }
-
-    public int getResult() {
-        Set<Integer> coordinates = cards.getCoordinateSums();
-        if (isBurst()) {
-            return getMinSum(coordinates);
-        }
-        return getMaxSum(coordinates);
-    }
-
-    private boolean isBurst() {
-        Set<Integer> coordinates = cards.getCoordinateSums();
-        return coordinates.stream().noneMatch(coordinate -> coordinate <= SUM_LIMIT);
-    }
-
-    private int getMaxSum(Set<Integer> coordinates) {
-        return coordinates.stream()
-                .filter(coordinate -> coordinate <= SUM_LIMIT)
-                .mapToInt(i -> i)
-                .max()
-                .orElseThrow(() -> new IllegalStateException("카드가 존재하지 않는 플레이어입니다."));
-    }
-
-    private int getMinSum(Set<Integer> coordinates) {
-        return coordinates.stream()
-                .mapToInt(i -> i)
-                .min()
-                .orElseThrow(() -> new IllegalStateException("카드가 존재하지 않는 플레이어입니다."));
     }
 }
